Total failure!
Although it does expand (1 + x)^n accurately if x and n are integers so the algorithm is sound, but that's not the point!
There's plenty of simpler methods to achieve that, the purpose is to expand for |x| < 1 and |n| < 1
The most probable cause of huge inaccuracy is the text stream interface.
The only viable solution seems to be to make it monolithic